Move base tests in test package
