+++ app-android/src/androidTest/kotlin/care/data4life/integration/app/test/BaseComposeTest.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2022 D4L data4life gGmbH - All rights reserved.
  */
 
-package care.data4life.integration.app.test.compose
+package care.data4life.integration.app.test
 
 import care.data4life.integration.app.MainActivity
 import care.data4life.integration.app.test.compose.junit5.createAndroidComposeExtension
