+++ app-android/src/androidTest/kotlin/care/data4life/integration/app/test/BaseSdkTest.kt
@@ -2,14 +2,14 @@
  * Copyright (c) 2022 D4L data4life gGmbH - All rights reserved.
  */
 
-package care.data4life.integration.app.crud
+package care.data4life.integration.app.test
 
 import care.data4life.integration.app.data.wrapper.Result
 import care.data4life.integration.app.data.wrapper.awaitLegacyListener
 import care.data4life.integration.app.page.HomePage
 import care.data4life.integration.app.page.onWelcomePage
 import care.data4life.integration.app.test.TestConfigLoader
-import care.data4life.integration.app.test.compose.BaseComposeTest
+import care.data4life.integration.app.test.BaseComposeTest
 import care.data4life.integration.app.test.compose.junit5.ComposeContentContext
 import care.data4life.sdk.Data4LifeClient
 import kotlinx.coroutines.runBlocking
