+++ app-android/src/androidTest/kotlin/care/data4life/integration/app/ui/common/atom/PrimaryTextButtonKtTest.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
-import care.data4life.integration.app.test.compose.BaseComposeTest
+import care.data4life.integration.app.test.BaseComposeTest
 import care.data4life.integration.app.test.compose.assertScreenshotMatches
 import care.data4life.integration.app.test.compose.setThemedContent
 import care.data4life.integration.app.test.compose.setThemedScreenshotContent
