+++ app-android/src/androidTest/kotlin/care/data4life/integration/app/ui/common/atom/TopLogoKtTest.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.unit.dp
-import care.data4life.integration.app.test.compose.BaseComposeTest
+import care.data4life.integration.app.test.BaseComposeTest
 import care.data4life.integration.app.test.compose.assertScreenshotMatches
 import care.data4life.integration.app.test.compose.setThemedContent
 import care.data4life.integration.app.test.compose.setThemedScreenshotContent
